requirements.txt
fastapi
sqlmodel
uvicorn[standard]
python-dotenv
httpx
psycopg2-binary
alembic
python-jose[cryptography]
python-multipart